+++ src/plugins/csp_handler/server/csp_handlers.ts
@@ -12,6 +12,7 @@ import {
   OnPreResponseToolkit,
   OpenSearchDashboardsRequest,
 } from '../../../core/server';
+import { parseCspHeader, stringifyCspHeader } from './csp_header_utils';
 
 const FRAME_ANCESTORS_DIRECTIVE = 'frame-ancestors';
 
@@ -94,25 +95,4 @@ function appendFrameAncestorsWhenMissing(
   }
 
   return updateNext(parsedCspHeader, "'self'", toolkit);
-}
-
-function parseCspHeader(cspHeader: string) {
-  const directives: string[] = cspHeader.split(';');
-
-  return directives.reduce((accumulator, directive) => {
-    const trimmed = directive.trim().split(' ');
-
-    accumulator.set(trimmed[0], trimmed.slice(1));
-
-    return accumulator;
-  }, new Map<string, string[]>());
-}
-
-function stringifyCspHeader(parsedCspHeader: Map<string, string[]>) {
-  const strings: string[] = [];
-  parsedCspHeader.forEach((values: string[], directive: string) => {
-    strings.push(directive + ' ' + values.join(' '));
-  });
-
-  return strings.join('; ');
 }
